[map-helper-functions] Create useColumnsRequiredChecker.ts hook
