+++ packages/map/src/helpers/index.ts
@@ -2,11 +2,9 @@ export { addUIDs } from './addUIDs'
 export { applyColorToLegend } from './applyColorToLegend'
 export { closeModal } from './closeModal'
 export { colorDistributions } from './colorDistributions'
-export { columnsRequiredChecker } from './columnsRequiredChecker'
 export { displayGeoName } from './displayGeoName'
 export { formatLegendLocation } from './formatLegendLocation'
 export { generateColorsArray } from './generateColorsArray'
-export { generateRuntimeLegend } from './generateRuntimeLegend'
 export { generateRuntimeLegendHash } from './generateRuntimeLegendHash'
 export { getGeoStrokeColor, getGeoFillColor } from './colors'
 export { getUniqueValues } from './getUniqueValues'
